Coalesce Product View refresh and asset ingestion
